+++ nodejs/databases/heroes/passives/h003.json
@@ -1,0 +1,7 @@
+{
+    "effects": [
+        {
+            "effect": "freeze"
+        }
+    ]   
+}
