+++ nodejs/import.sh
@@ -1,0 +1,3 @@
+#!/bin/sh
+
+node src/importdb.ts
