+++ nodejs/src/types/zones.ts
@@ -1,0 +1,23 @@
+
+export type Zone = 
+| "bf1"
+| "bf2"
+| "bf3"
+| "bf4"
+| "bf5"
+| "bf6"
+| "bf7"
+| "bf8"
+| "hand"
+| "library"
+| "graveyard";
+
+export type BfZone = 
+| "bf1"
+| "bf2"
+| "bf3"
+| "bf4"
+| "bf5"
+| "bf6"
+| "bf7"
+| "bf8";
